+++ src/terrain.rs
@@ -1,9 +1,10 @@
 use crate::block::Block;
 use crate::mesh::mesh::MeshType;
+use crate::texture::TextureAtlas;
 
 use std::collections::HashMap;
-use crate::texture::TextureAtlas;
 use std::rc::Rc;
+use rand::Rng;
 
 enum Biome {
     FlatPlains
@@ -17,9 +18,10 @@ impl Terrain {
     pub fn new(txtr: Rc<TextureAtlas>) -> Self {
         let mut blockspace = HashMap::new();
 
-        blockspace.insert("air", Block::new(MeshType::Cube, txtr.clone(), "air", &[[3,0], [3,0], [3,0], [3, 0], [3,0], [3,0]], 0));
-        blockspace.insert("dirt", Block::new(MeshType::Cube, txtr.clone(), "dirt", &[[2,0], [2,0], [2,0], [2, 0], [2,0], [2,0]], 0));
-        blockspace.insert("grass", Block::new(MeshType::Cube, txtr.clone(), "grass", &[[0,0], [2,0], [1,0], [1,0], [1,0], [1,0]], 0));
+        blockspace.insert("air", Block::new(MeshType::Cube, txtr.clone(), "air", &[[4,0], [4,0], [4,0], [4, 0], [4,0], [4,0]], 0, true));
+        blockspace.insert("dirt", Block::new(MeshType::Cube, txtr.clone(), "dirt", &[[2,0], [2,0], [2,0], [2, 0], [2,0], [2,0]], 0, false));
+        blockspace.insert("grass", Block::new(MeshType::Cube, txtr.clone(), "grass", &[[0,0], [2,0], [1,0], [1,0], [1,0], [1,0]], 0, false));
+        blockspace.insert("stone", Block::new(MeshType::Cube, txtr.clone(), "stone", &[[3,0], [3,0], [3,0], [3, 0], [3,0], [3,0]], 0, false));
 
         Self {
             blocks: blockspace,
@@ -33,12 +35,17 @@ impl Terrain {
         let mut block_data: Vec<Block> = Vec::with_capacity(size*size*size);
 
         for x in position[0]..position[0]+size as u32 {
+            let num = rand::thread_rng().gen_range(0, 3);
             for y in position[1]..position[1]+size as u32 {
                 for z in position[2]..position[2]+size as u32  {
-                    if y < size as u32-1 {
+                    if y >= 32-num as u32-1 {
+                        block_data.push(self.blocks["air"].clone());
+                    } else if y >= 32-num-2 {
+                        block_data.push(self.blocks["grass"].clone());
+                    } else if y >= 32-num-5 {
                         block_data.push(self.blocks["dirt"].clone());
                     } else {
-                        block_data.push(self.blocks["grass"].clone());
+                        block_data.push(self.blocks["stone"].clone());
                     }
                 }
             }